+++ Frontend/package.json
@@ -11,6 +11,7 @@
     "date-fns": "^2.30.0",
     "http-proxy-middleware": "^2.0.6",
     "react": "^18.2.0",
+    "react-date-range": "^1.4.0",
     "react-datepicker": "^4.16.0",
     "react-dom": "^18.2.0",
     "react-icons": "^4.10.1",
